requirements.txt
Django==5.0.6
django_select2==8.2.0
djangorestframework==3.15.2
